+++ priv/templates/brando.install/web/controllers/lockdown_controller.ex
@@ -6,4 +6,18 @@ defmodule <%= application_module %>.LockdownController do
     |> put_layout({<%= application_module %>.LayoutView, "lockdown.html"})
     |> render("index.html")
   end
+
+  def post_password(conn, %{"password" => password}) do
+    hashed_pass = Comeonin.Bcrypt.hashpwsalt(Brando.config(:lockdown_password))
+
+    if Comeonin.Bcrypt.checkpw(password, hashed_pass) do
+      conn
+      |> fetch_session
+      |> put_session(:lockdown_authorized, true)
+      |> redirect(to: "/")
+    else
+      conn
+      |> redirect(to: Brando.helpers.lockdown_path(conn, :index))
+    end
+  end
 end