+++ test/brando/auth/controller_test.exs
@@ -5,8 +5,17 @@ defmodule Brando.Auth.ControllerTest do
   use Plug.Test
   use RouterHelper
 
-  @login %{"email" => "james@thestooges.com", "password" => "hunter2hunter2"}
-  @bad_login %{"email" => "bad@gmail.com", "password" => "finimeze"}
+  alias Brando.Factory
+
+  @login %{
+    "email" => "james@thestooges.com",
+    "password" => "hunter2hunter2"
+  }
+
+  @bad_login %{
+    "email" => "bad@gmail.com",
+    "password" => "finimeze"
+  }
 
   test "login get" do
     conn =
@@ -18,34 +27,39 @@ defmodule Brando.Auth.ControllerTest do
   end
 
   test "login post ok" do
-    Forge.saved_user_w_hashed_pass(TestRepo)
+    Factory.create(:user)
+
     conn =
       :post
       |> call("/login", %{"user" => @login})
       |> with_session
       |> send_request
+
     assert redirected_to(conn, 302) =~ "/admin"
   end
 
   test "login post failed" do
-    Forge.saved_user_w_hashed_pass(TestRepo)
+    Factory.create(:user)
+
     conn =
       :post
       |> call("/login", %{"user" => @bad_login})
       |> with_session
       |> send_request
+
     assert redirected_to(conn, 302) =~ "/login"
     assert get_flash(conn, :error) == "Authorization failed"
   end
 
   test "logout" do
-    user = Forge.saved_user_w_hashed_pass(TestRepo)
+    user = Factory.create(:user)
+
     conn =
       :get
       |> call("/logout")
       |> with_user(user)
       |> send_request
-    assert html_response(conn, 200)
-           =~ "You have been logged out of the admin area"
+
+    assert html_response(conn, 200) =~ "You have been logged out of the admin area"
   end
 end