+++ test/test_helper.exs
@@ -1,4 +1,7 @@
 Logger.configure(level: :info)
+
+{:ok, _} = Application.ensure_all_started(:ex_machina)
+
 ExUnit.start
 
 # Clear tmp dir
@@ -43,7 +46,6 @@ defmodule Brando.Integration.TestCase do
 end
 
 Code.require_file "support/user_socket.exs", __DIR__
-Code.require_file "support/fixtures.exs", __DIR__
 
 _ = Ecto.Storage.down(Repo)
 _ = Ecto.Storage.up(Repo)